LayoutTests/platform/glib/imported/w3c/web-platform-tests/xhr/access-control-basic-non-cors-safelisted-content-type-expected.txt

CONSOLE MESSAGE: Request header field Content-Type is not allowed by Access-Control-Allow-Headers.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.web-platform.test:8800/xhr/resources/access-control-basic-put-allow.py due to access control checks.

PASS Deny cross-origin request with non-CORS-safelisted content type
